+++ api/tests/test_permissions.py
@@ -8,12 +8,12 @@ class RestFrameworkPermissionsTest(TestCase):
         client = RequestsClient()
         client.headers.update({'x-test': 'true'})
 
-        response = client.get('http://testserver/documents/')
+        response = client.get('http://testserver/document/')
         assert response.status_code == 403
 
     def test_options_request_works_without_auth(self):
         client = RequestsClient()
         client.headers.update({'x-test': 'true'})
 
-        response = client.options('http://testserver/documents/')
+        response = client.options('http://testserver/document/')
         assert response.status_code == 200
