+++ document_service/wsgi.py
@@ -1,5 +1,5 @@
 """
-WSGI config for documents project.
+WSGI config for document project.
 
 It exposes the WSGI callable as a module-level variable named ``application``.
 
@@ -12,6 +12,6 @@ import os
 from django.core.wsgi import get_wsgi_application
 
 os.environ.setdefault("DJANGO_SETTINGS_MODULE",
-                      "documents-service.settings.production")
+                      "document_service.settings.production")
 
 application = get_wsgi_application()
