+++ manage.py
@@ -4,7 +4,7 @@ import sys
 
 if __name__ == "__main__":
     os.environ.setdefault("DJANGO_SETTINGS_MODULE",
-                          "documents-service.settings.base")
+                          "document_service.settings.base")
     try:
         from django.core.management import execute_from_command_line
     except ImportError as exc:
